app/src/main/java/com/sreeni/testsamples/RegistrationUtil.kt
package com.sreeni.testsamples

object RegistrationUtil {

    val usersList = listOf("Sreenu", "Ramu")

    fun validateRegistrationForm(
        username: String,
        password: String,
        confirmPassword: String
    ): Boolean {
        if (username.isEmpty() || password.isEmpty()) {
            return false
        }
        if (username in usersList) {
            return false
        }
        if (password != confirmPassword) {
            return false
        }
        if (password.count { it.isDigit() } < 2) {
            return false
        }
        return true
    }

    fun checkBraces(string: String): Boolean {
        return string.count { it == '(' } == string.count { it == ')' }
    }
}